Added authentication and displaying data for particular gym

+++ lib/Screens/MemberListPage.dart
@@ -4,6 +4,7 @@ import 'package:arjunagym/Screens/MemberDetails.dart';
 import 'package:arjunagym/Screens/SearchScreen.dart';
 import 'package:arjunagym/Screens/renewEdit.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:arjunagym/Models/GymPlan.dart';
@@ -192,20 +193,31 @@ class MemberListSection extends StatelessWidget {
 
                 if (confirmDelete == true) {
                   try {
-                    await FirebaseFirestore.instance
-                        .collection('members')
-                        .doc(member.id)
-                        .delete();
-                    await FirebaseFirestore.instance
-                        .collection('recyclebin')
-                        .doc(member.id)
-                        .set(member.toMap());
+                    User? currentUser = FirebaseAuth.instance.currentUser;
+                    if (currentUser != null) {
+                      // Delete the member from the user's members subcollection
+                      await FirebaseFirestore.instance
+                          .collection('Users')
+                          .doc(currentUser.uid)
+                          .collection('members')
+                          .doc(member.id)
+                          .delete();
 
-                    ScaffoldMessenger.of(context).showSnackBar(
-                      SnackBar(content: Text('Member deleted successfully')),
-                    );
+                      // Move the member to the user's recyclebin subcollection
+                      await FirebaseFirestore.instance
+                          .collection('Users')
+                          .doc(currentUser.uid)
+                          .collection('recyclebin')
+                          .doc(member.id)
+                          .set(member.toMap());
 
-                    memberProvider.getAllMembers();
+                      ScaffoldMessenger.of(context).showSnackBar(
+                        SnackBar(content: Text('Member deleted successfully')),
+                      );
+
+                      // Refresh the member list
+                      Provider.of<MemberProvider>(context, listen: false).getAllMembers();
+                    }
                   } catch (error) {
                     print('Error deleting member: $error');
                     ScaffoldMessenger.of(context).showSnackBar(
